Day 6 or 7 (sorry lmao)
In Day 6 part 1 the hardest thing was actually to parse the input
and read the numbers through all these weird spaces. :D But in part 2
I just treated it as char** and clutch idea to read numbers vertically
was to copy chars into second string and atoi() it.

Day 7 part 1 was just doing 'visually' the beam movement on char** array
and in part 2 it was reused and then just recursive function
implemented, but it quickly shown that results need to be hashed. Sadly
I have lost much time debugging "too low answer" when I was pretty sure
algorithm was correct only to read in some post that result was too big
for an int value. :(