Day 8
The hardest part was to debug why result is wrong, even though I was
almost sure my algorithm is correct. Root cause was... sorting function
that wasn't working correctly due to 'bug' in comparision function
which was doing simple 'classic' `return a-b` thing, but it turns out
that it doesn't work for long ints, and after switching to if's the
result suddenly is just correct.

Part 2 with so solid structure from Part 1 was matter of less than 1
minute to achieve.